quick upload

+++ src/lib/computing/algorithms/recursion/factorial/Factorial.java
@@ -3,8 +3,8 @@ package lib.computing.algorithms.recursion.factorial;
 import java.text.NumberFormat;
 
 public final class Factorial {
-    private int number;
-    private int factorial(int number) {
+    private long number;
+    private long factorial(long number) {
         if (number != 0) {
             return number * factorial(number - 1);
         }
@@ -13,7 +13,7 @@ public final class Factorial {
         }
     }
     public void getFactorial() {
-        int result = factorial(number);
+        long result = factorial(number);
         if (number == 0) {
             System.out.println("The Factorial of " + this.number + " is: " + 0);
         }
@@ -21,7 +21,7 @@ public final class Factorial {
             System.out.println("The Factorial of " + this.number + " is: " + NumberFormat.getInstance().format(result));
         }
     }
-    public Factorial setNumber(int number) {
+    public Factorial setNumber(long number) {
         this.number = number;
         return this;
     }
